fix(packaging): make the reg.exe timeout real, and stop inferring "absent" from an exit code
Review round 6 on #49 — both findings are this file's own stated guarantees not holding.

- The 5s timeout did not bound the call: `readText()` on the child's stdout blocks until EOF,
  which for a wedged reg.exe (registry lock, EDR hook, filter driver) is when it exits, so
  `waitFor(5s)` was only ever reached afterwards. The KDoc promised an uninstaller waiting on
  `--unregister-protocol` could not be hung by that, and it could. Output is now redirected to
  a temp file, so `waitFor` runs first and the budget is real — the naive fix (swapping the two
  lines) would trade the hang for a pipe-buffer deadlock, which is why the reasoning is in the
  KDoc.
- queryRootKeyPresent inferred absence from a bare exit code, which is exactly the conflation
  parseCommandState exists to avoid: an access-denied or policy-blocked `reg query` runs fine
  and exits 1, so it reported rootPresent=false -> Report(ABSENT) -> exit 0, telling an
  uninstaller nothing was left to clean while a live registration remained and nothing had
  been checked. Now a pure parseRootKeyPresence(exitCode, output) with the same three-state
  rule as the command parse, plus 4 tests.

The pure decision layer also moves to its own file, WindowsProtocolCleanupPolicy.kt: it is the
same split as DebControl (registry needs Windows, policy does not, policy is where the
mistakes are), and it keeps both files under detekt's function-count thresholds. 30 tests in
WindowsProtocolCleanupTest.
